requirements.txt
flask>=3.0
requests>=2.32
beautifulsoup4>=4.12
httpx>=0.27
python-dotenv>=1.0
